moved all integers to defined macros

+++ source/string.h
@@ -3,8 +3,8 @@
 
 struct string {
 	char* data;
-	size_t cap;
-	size_t len;
+	i32 cap;
+	i32 len;
 };
 
 bool string_from_file( struct string* source, char* filename ){
@@ -12,7 +12,7 @@ bool string_from_file( struct string* source, char* filename ){
 	assert( source->cap == 0 && source->len == 0 && source->data == NULL, "String must be empty" );
 	assert( filename != NULL, "Malformed args" );
 	struct stat stat_buffer = { 0 };
-	int stat_err = stat( filename, &stat_buffer );
+	i32 stat_err = stat( filename, &stat_buffer );
 	if( stat_err == -1 ){
 		char* tmp = malloc( 1 );
 		assert( tmp != NULL, "Alloc failed." );
@@ -29,8 +29,8 @@ bool string_from_file( struct string* source, char* filename ){
 		char* tmp = malloc( stat_buffer.st_size + 1 );
 		assert( tmp != NULL, "Alloc failed." );
 		source->data = tmp;
-		size_t fread_res = fread( source->data, 1, stat_buffer.st_size, file );
-		assert( (size_t) stat_buffer.st_size == fread_res, "fread failed" );
+		i64 fread_res = (i64) fread( source->data, 1, stat_buffer.st_size, file );
+		assert( stat_buffer.st_size == fread_res, "fread failed" );
 		fclose( file );
 		source->cap = stat_buffer.st_size + 1;
 		source->len = stat_buffer.st_size ;
@@ -44,13 +44,13 @@ bool string_to_file( struct string* source, char* filename ){
 	assert( source->cap > source->len || source->len <= 0, "Malformed internal source data" );
 	assert( filename != NULL, "Malformed args" );
 	struct stat stat_buffer = { 0 };
-	int stat_err = stat( filename, &stat_buffer );
+	i32 stat_err = stat( filename, &stat_buffer );
 	if( stat_err != -1 && !S_ISREG( stat_buffer.st_mode )){
 		return true;
 	}
 	FILE* file = fopen( filename, "w" );
 	assert( file != NULL, "fopen failed" );
-	size_t fwrite_res = fwrite( source->data, 1, source->len, file );
+	i64 fwrite_res = (i64) fwrite( source->data, 1, source->len, file );
 	assert( source->len == fwrite_res, "fwrite failed" );
 	fclose( file );
 	return false;
@@ -66,13 +66,13 @@ void string_free( struct string* source ){
 	source->len = 0;
 }
 
-void string_append( struct string* source, char* src, size_t count ){
+void string_append( struct string* source, char* src, i32 count ){
 	assert( source != NULL, "Malformed args" );
 	assert( src != NULL, "Malformed args" );
 	assert( source->cap > source->len || source->len <= 0, "Malformed internal source data" );
 	if( count == 0 ) return;
 	if( source->len + count >= source->cap ){
-		size_t cap = ( source->cap + count ) * 2;
+		i32 cap = ( source->cap + count ) * 2;
 		char* tmp = realloc( source->data, cap );
 		assert( tmp != NULL, "Alloc failed" );
 		source->data = tmp;
